+++ commands/CAccount.js
@@ -1083,18 +1083,18 @@ class CAccount {
                         command: "application metrics <appid>",
                         details: "List application's metrics"
                     },
-                    {
-                        level: "app_admin",
-                        theme: " ",
-                        command: " ",
-                        details: " "
-                    },
-                    {
-                        level: "app_admin",
-                        theme: "Analytics",
-                        command: "application analytics <appid> <since> <until>",
-                        details: "List application's analytics"
-                    },
+                    // {
+                    //     level: "app_admin",
+                    //     theme: " ",
+                    //     command: " ",
+                    //     details: " "
+                    // },
+                    // {
+                    //     level: "app_admin",
+                    //     theme: "Analytics",
+                    //     command: "application analytics <appid> <since> <until>",
+                    //     details: "List application's analytics"
+                    // },
                     {
                         level: "app_admin",
                         theme: " ",